+++ src/main/java/com/barion/dungeons_enhanced/DungeonsEnhanced.java
@@ -9,31 +9,28 @@ import com.barion.dungeons_enhanced.world.DETemplatePools;
 import com.barion.dungeons_enhanced.world.structure.processor.DEProcessors;
 import com.google.common.reflect.Reflection;
 import com.legacy.structure_gel.api.registry.registrar.RegistrarHandler;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.data.event.GatherDataEvent;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.config.ModConfig;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import net.minecraft.core.RegistrySetBuilder;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.ModLoadingContext;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.config.ModConfig;
+import net.neoforged.neoforge.common.data.DatapackBuiltinEntriesProvider;
+import net.neoforged.neoforge.data.event.GatherDataEvent;
+
+import java.util.Set;
 
 @Mod(DungeonsEnhanced.MOD_ID)
 public class DungeonsEnhanced{
     public static final String MOD_ID = "dungeons_enhanced";
-    public static final Logger LOGGER = LogManager.getLogger();
 
-    public DungeonsEnhanced() {
+    public DungeonsEnhanced(IEventBus modEventBus) {
         ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, DEConfig.COMMON_SPEC);
-        final IEventBus modBus = FMLJavaModLoadingContext.get().getModEventBus();
-        final IEventBus forgeBus = MinecraftForge.EVENT_BUS;
         Reflection.initialize(DEStructures.class, DEProcessors.class, DETemplatePools.class);
 
-        modBus.addListener(DungeonsEnhanced::gatherData);
-        modBus.addListener(DEJigsawTypes::register);
+        modEventBus.addListener(DungeonsEnhanced::gatherData);
+        modEventBus.addListener(DEJigsawTypes::register);
 
-        RegistrarHandler.registerHandlers(MOD_ID, modBus, DETemplatePools.HANDLER, DEProcessors.HANDLER);
+        RegistrarHandler.registerHandlers(MOD_ID, modEventBus, DETemplatePools.HANDLER, DEProcessors.HANDLER);
     }
 
     public static void gatherData(GatherDataEvent event){
@@ -42,9 +39,9 @@ public class DungeonsEnhanced{
         var existingFileHelper = event.getExistingFileHelper();
         var runServer = event.includeServer();
 
-        var registrarProvider = RegistrarHandler.createGenerator(output, MOD_ID);
-        generator.addProvider(event.includeServer(), registrarProvider);
-        var lookup = registrarProvider.getLookupProvider();
+        var registrarProvider = new DatapackBuiltinEntriesProvider(output, event.getLookupProvider(), RegistrarHandler.injectRegistries(new RegistrySetBuilder()), Set.of(DungeonsEnhanced.MOD_ID));
+        generator.addProvider(runServer, registrarProvider);
+        var lookup = registrarProvider.getRegistryProvider();
 
         generator.addProvider(runServer, new DEBiomeTagsProvider(output, event.getLookupProvider(), existingFileHelper));
         generator.addProvider(runServer, new DELootTableProvider(output));