+++ src/main/java/com/barion/dungeons_enhanced/data/provider/DEStructureTagsProvider.java
@@ -12,7 +12,7 @@ import net.minecraft.data.tags.StructureTagsProvider;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.levelgen.structure.Structure;
-import net.minecraftforge.common.data.ExistingFileHelper;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;