+++ src/main/java/com/barion/dungeons_enhanced/world/DEJigsawTypes.java
@@ -1,30 +1,19 @@
 package com.barion.dungeons_enhanced.world;
 
-import com.barion.dungeons_enhanced.DEUtil;
+import com.barion.dungeons_enhanced.DungeonsEnhanced;
 import com.barion.dungeons_enhanced.world.structure.*;
-import com.legacy.structure_gel.api.events.RegisterJigsawCapabilityEvent;
-import com.legacy.structure_gel.api.structure.jigsaw.JigsawCapability;
+import com.legacy.structure_gel.api.registry.StructureGelRegistries;
+import com.legacy.structure_gel.api.registry.registrar.Registrar;
+import com.legacy.structure_gel.api.registry.registrar.RegistrarHandler;
+import com.legacy.structure_gel.api.structure.jigsaw.JigsawCapabilityType;
 
 public class DEJigsawTypes {
-    public static final JigsawCapability.JigsawType<DECastle.Capability> CASTLE = ()-> DECastle.Capability.CODEC;
-    public static final JigsawCapability.JigsawType<DEDeepCrypt.Capability> DEEP_CRYPT = ()-> DEDeepCrypt.Capability.CODEC;
-    public static final JigsawCapability.JigsawType<DEDesertTomb.Capability> DESERT_TOMB = ()-> DEDesertTomb.Capability.CODEC;
-    public static final JigsawCapability.JigsawType<DEDruidCircle.Capability> DRUID_CIRCLE = ()-> DEDruidCircle.Capability.CODEC;
-    public static final JigsawCapability.JigsawType<DELargeDungeon.Capability> LARGE_DUNGEON = ()-> DELargeDungeon.Capability.CODEC;
-    public static final JigsawCapability.JigsawType<DEMonsterMaze.Capability> MONSTER_MAZE = ()-> DEMonsterMaze.Capability.CODEC;
-    public static final JigsawCapability.JigsawType<DEPillagerCamp.Capability> PILLAGER_CAMP = ()-> DEPillagerCamp.Capability.CODEC;
-
-    public static void register(final RegisterJigsawCapabilityEvent event){
-        register(event, "castle", CASTLE);
-        register(event, "deep_crypt", DEEP_CRYPT);
-        register(event, "desert_tomb", DESERT_TOMB);
-        register(event, "druid_circle", DRUID_CIRCLE);
-        register(event, "large_dungeon", LARGE_DUNGEON);
-        register(event, "monster_maze", MONSTER_MAZE);
-        register(event, "pillager_camp", PILLAGER_CAMP);
-    }
-
-    private static void register(final RegisterJigsawCapabilityEvent event, String key, JigsawCapability.JigsawType<?> type){
-        event.register(DEUtil.location(key), type);
-    }
+    public static final RegistrarHandler<JigsawCapabilityType<?>> HANDLER = RegistrarHandler.getOrCreate(StructureGelRegistries.Keys.JIGSAW_TYPE, DungeonsEnhanced.MOD_ID);
+    public static final Registrar.Static<JigsawCapabilityType<DECastle.Capability>> CASTLE = HANDLER.createStatic(DECastle.ID, () -> () -> DECastle.Capability.CODEC);
+    public static final Registrar.Static<JigsawCapabilityType<DEDeepCrypt.Capability>> DEEP_CRYPT = HANDLER.createStatic(DEDeepCrypt.ID, () -> () -> DEDeepCrypt.Capability.CODEC);
+    public static final Registrar.Static<JigsawCapabilityType<DEDesertTomb.Capability>> DESERT_TOMB = HANDLER.createStatic(DEDesertTomb.ID, () -> () -> DEDesertTomb.Capability.CODEC);
+    public static final Registrar.Static<JigsawCapabilityType<DEDruidCircle.Capability>> DRUID_CIRCLE = HANDLER.createStatic(DEDruidCircle.ID, () -> () -> DEDruidCircle.Capability.CODEC);
+    public static final Registrar.Static<JigsawCapabilityType<DELargeDungeon.Capability>> LARGE_DUNGEON = HANDLER.createStatic(DELargeDungeon.ID, () -> () -> DELargeDungeon.Capability.CODEC);
+    public static final Registrar.Static<JigsawCapabilityType<DEMonsterMaze.Capability>> MONSTER_MAZE = HANDLER.createStatic(DEMonsterMaze.ID, () -> () -> DEMonsterMaze.Capability.CODEC);
+    public static final Registrar.Static<JigsawCapabilityType<DEPillagerCamp.Capability>> PILLAGER_CAMP = HANDLER.createStatic(DEPillagerCamp.ID, () -> () -> DEPillagerCamp.Capability.CODEC);
 }