+++ src/main/java/com/barion/dungeons_enhanced/world/structure/DECastle.java
@@ -4,10 +4,7 @@ import com.barion.dungeons_enhanced.DEStructures;
 import com.barion.dungeons_enhanced.DungeonsEnhanced;
 import com.barion.dungeons_enhanced.world.DEJigsawTypes;
 import com.barion.dungeons_enhanced.world.DETemplatePools;
-import com.legacy.structure_gel.api.structure.jigsaw.ExtendedJigsawStructurePiece;
-import com.legacy.structure_gel.api.structure.jigsaw.IPieceFactory;
-import com.legacy.structure_gel.api.structure.jigsaw.JigsawCapability;
-import com.legacy.structure_gel.api.structure.jigsaw.JigsawRegistryHelper;
+import com.legacy.structure_gel.api.structure.jigsaw.*;
 import com.mojang.serialization.Codec;
 import net.minecraft.core.BlockPos;
 import net.minecraft.data.worldgen.BootstapContext;
@@ -22,13 +19,12 @@ import net.minecraft.world.level.levelgen.structure.pools.StructureTemplatePool;
 public class DECastle {
     public static final String ID = "castle";
 
-    public static class Capability implements JigsawCapability.IJigsawCapability{
-        public static final Capability Instance = new Capability();
-        public static final Codec<Capability> CODEC = Codec.unit(Instance);
+    public static class Capability implements JigsawCapability{
+        public static final Capability INSTANCE = new Capability();
+        public static final Codec<Capability> CODEC = Codec.unit(INSTANCE);
 
         @Override
-        public JigsawCapability.JigsawType<?> getType() {return DEJigsawTypes.CASTLE;}
-
+        public JigsawCapabilityType<?> getType() {return DEJigsawTypes.CASTLE.get();}
         @Override
         public IPieceFactory getPieceFactory() {return Piece::new;}