+++ src/main/java/com/barion/dungeons_enhanced/world/structure/DEDeepCrypt.java
@@ -19,12 +19,12 @@ import net.minecraft.world.level.levelgen.structure.pools.StructureTemplatePool;
 
 public class DEDeepCrypt{
     public static final String ID = "deep_crypt";
-    public static class Capability implements JigsawCapability.IJigsawCapability{
-        public static final Capability Instance = new Capability();
-        public static final Codec<Capability> CODEC = Codec.unit(Instance);
+    public static class Capability implements JigsawCapability{
+        public static final Capability INSTANCE = new Capability();
+        public static final Codec<Capability> CODEC = Codec.unit(INSTANCE);
 
         @Override
-        public JigsawCapability.JigsawType<?> getType(){return DEJigsawTypes.DEEP_CRYPT;}
+        public JigsawCapabilityType<?> getType(){return DEJigsawTypes.DEEP_CRYPT.get();}
         @Override
         public IPieceFactory getPieceFactory() {return Piece::new;}
     }